Lay groundwork for new design with website lists being autogened
Apparently, Big Tech websites have a lot of domains associated with
themselves. So this is a move away from user-defined domains and towards
using a tool in this repo (website-domain-gen) to take a blocklist and a
TOML file as input and spit out a TOML database of websites and their
associated IP addresses.

This commit starts that process by revamping the config structures in
config.rs, updating the TOML file definition, and fixing all the
breakages that causes.

Signed-off-by: Tristan Andrus <tristan@tristanandrus.com>